% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_with_rot.R
\name{axis_label}
\alias{axis_label}
\title{Add axis labels with rotation}
\usage{
axis_label(
  side,
  at,
  labels = at,
  type = "long",
  line = NA,
  srt = NA,
  font = NA,
  col = NULL,
  hadj = NA,
  padj = NA,
  ...
)
}
\arguments{
\item{side}{An integer specifying which side of the plot
the axis is to be shown.
Equivalent to \code{side} argument in \link{axis}.}

\item{at}{The points at which labels are to be drawn.
Passing values of \link{axis} is possible.}

\item{labels}{Equivalent to \code{labels} argument in \link{axis} function.}

\item{type}{An integer either \code{'long'}, \code{'short'}
or their abbreviations. It determines the center position of labels
when \code{hadj} and \code{padj} are not given.}

\item{srt}{An angle for label rotation (string rotation).}

\item{font}{An integer indicating font of labels.}

\item{col}{An integer or a string indicating color of labels.}

\item{hadj}{An integer indicating adjustment (see par("adj"))
for all labels parallel (‘horizontal’) to the reading direction.}

\item{padj}{An integer indicating adjustment (see par("adj"))
for all labels perpendicular to the reading direction.}
}
\description{
Add axis labels with rotation
}
\examples{
data('iris')
b <- boxplot(Sepal.Length ~ Species, data = iris, xaxt = 'n')
at <- axis(side = 1, at = seq_along(b$names), labels = FALSE)
axis_label(side = 1, at = at, labels = b$names, srt = 45)

}
