% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_on_baseplot.R
\name{colorbar}
\alias{colorbar}
\title{カラーバーを描画する関数}
\usage{
colorbar(
  x,
  y,
  width,
  height,
  color,
  limits,
  at,
  n_color = 50,
  title = "Value",
  horizontal = TRUE,
  lwd_bar = 1,
  border_color = "black",
  tick_len = 0.02,
  lwd_ticks = 1,
  show_label = TRUE,
  cex_label = 1,
  space_height_label = 0.02,
  side_title = "left",
  space_width_title = 0.03,
  space_height_title = 0.03,
  cex_title = 1
)
}
\description{
カラーバーを描画する関数
}
