% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif_mark.R
\name{convert_pval_to_txt}
\alias{convert_pval_to_txt}
\title{Convert p-value to significance mark}
\usage{
convert_pval_to_txt(p)
}
\arguments{
\item{p}{A numeric vector of p-values.}
}
\value{
Characters either \code{"n.s."}, \code{"*"}, \code{"**"}, or \code{"***"},
    if \eqn{p > 0.05}, \eqn{p <= 0.05}, \eqn{p <= 0.01}, or \eqn{p <= 0.001},
    respectively.
}
\description{
Convert p-value to significance mark
}
