% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_barplot.R
\name{easy_barplot}
\alias{easy_barplot}
\title{Easy making of barplot}
\usage{
easy_barplot(
  y,
  x1,
  x2 = NULL,
  test_method = NULL,
  signif_label_at_top = TRUE,
  signif_label_size = 2,
  signif_label_space = 0.02,
  show_errorbar = TRUE,
  errorbar_shift = -0.2,
  errorbar_length = 0.1,
  show_point = TRUE,
  point_shift = NA,
  point_color = "gray20",
  point_bg = "white",
  point_cex = 1,
  point_pch = 1,
  point_lwd = 1,
  axislabel_x2_line = 3,
  y_space = 0.03,
  main = ""
)
}
\arguments{
\item{y}{A numeric vector.}

\item{x1}{A factor vector used for grouping \code{y}.
The length should be the same as \code{y}.}

\item{x2}{An optional factor vector used for grouping bars.
The length should be the same as \code{y}.}

\item{test_method}{A character determining how the differences among groups
should be tested. \code{'tukey'} or \code{'dunnett'} can be given.
If \code{NULL}, no test will be applied.
If \code{x2} is provided, the tests will be applied for
each groups of \code{x2} independently.}

\item{signif_label_at_top}{Logical. Should the labels shown at the top of the bars?}

\item{signif_label_size}{Label size.}

\item{signif_label_space}{Space between labels and any nearest object
(either of bars, error bars, or points).}

\item{show_errorbar}{Logical. Should error bars be shown?}

\item{errorbar_shift}{How much x coordinates of the error bars
should be shifted from the center of the bars.
It will be set at \code{-0.2} if \code{show_point} is \code{TRUE}, otherwise \code{0}.}

\item{errorbar_length}{Length of horizontal lines of error bars.}

\item{show_point}{Logical. Should the data points be shown?}

\item{point_shift}{How much x coordinates of the points
should be shifted from the center of the bars.
It will be set at \code{0.2} if \code{show_errorbar} is \code{TRUE}, otherwise \code{0}.}

\item{point_color}{Color of points.}

\item{point_bg}{Background color of points (applied if \code{point_pch} is in 21-25).}

\item{point_cex}{Point size.}

\item{point_pch}{Point shape.}

\item{point_lwd}{Line width of points.}
}
\description{
This function is a wrapper of \code{\link[graphics]{barplot}} function.
}
