% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_on_baseplot.R
\name{heatmap_on_baseplot}
\alias{heatmap_on_baseplot}
\title{Heatmap on base plot}
\usage{
heatmap_on_baseplot(
  x,
  row_group,
  col_group,
  scale = "none",
  color = rcolorb,
  margins = c(5, 5),
  cell_height = 1,
  cell_width = 1,
  sep_height = 0,
  sep_width = 0,
  groupsep_height = 0.5,
  groupsep_width = 0.5,
  cex_row = NULL,
  cex_col = NULL,
  main = "",
  xlab = "",
  ylab = ""
)
}
\description{
Heatmap on base plot
}
