% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{legend_outside}
\alias{legend_outside}
\title{legend() outside plot}
\usage{
legend_outside(...)
}
\description{
This equals to `legend(x = par()$usr[2], y = par()$usr[4], box.col = NA,
xjust   = 1, yjust   = 0, ...)`.
}
