% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matplot.R
\name{matplot0}
\alias{matplot0}
\title{Expanded matplot()}
\usage{
matplot0(
  x = NULL,
  y,
  type = "l",
  lty = 1:5,
  lwd = 1,
  lend = par("lend"),
  pch = 1,
  col = 1:6,
  cex = 1,
  bg = 1,
  xlab = "x",
  ylab = "y",
  xlim = NULL,
  ylim = NULL
)
}
\description{
Expanded matplot()
}
\details{
This function is very similar to `matplot()`,
but removes NA when plotting.
}
