% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot_layout.R
\name{multiplot_layout}
\alias{multiplot_layout}
\title{Segmenting plot area beautifully}
\usage{
multiplot_layout(
  mat,
  upper_space = 0.05,
  lower_space = 0.05,
  right_space = 0.05,
  left_space = 0.05,
  ...
)
}
\description{
Segmenting plot area beautifully
}
