% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coord.R
\name{plot_coord_x}
\alias{plot_coord_x}
\title{Convert relative coordinates to absolute coordinates in plot region}
\usage{
plot_coord_x(x)
}
\description{
Convert relative coordinates to absolute coordinates in plot region
}
