% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coord.R
\name{plot_coord_y}
\alias{plot_coord_y}
\title{Convert relative coordinates to absolute coordinates in plot region}
\usage{
plot_coord_y(y)
}
\description{
Convert relative coordinates to absolute coordinates in plot region
}
