% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsupporter-package.R
\docType{package}
\name{plotsupporter-package}
\alias{plotsupporter}
\alias{plotsupporter-package}
\title{plotsupporter: A package of convenient functions supporting base plot}
\description{
Convenient functions for visualization were made for myself.
}
\keyword{internal}
