% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_with_rot.R
\name{xaxis_with_vert_label}
\alias{xaxis_with_vert_label}
\title{Convert p-value to significance mark}
\usage{
xaxis_with_vert_label(at, labels, srt = 90, pos = 0.05, adj = c(1, 0.5))
}
\arguments{
\item{at}{Equivalent to `at` argument in `axis()` function.}

\item{labels}{Equivalent to `labels` argument in `axis()` function.}

\item{srt}{An angle for label rotation.}

\item{pos}{A parameter to adjust vertical positions of labels.}

\item{adj}{A parameter to adjust relationship between coordinates of labels
and actual positions of labels.}
}
\description{
Convert p-value to significance mark
}
