% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_custom.R
\name{colorbar}
\alias{colorbar}
\title{カラーバーを描画する関数}
\usage{
colorbar(
  x,
  y,
  bar_col = cm.colors(7),
  n_bar_col = 30,
  horizontal = TRUE,
  bar_width = if (horizontal) 5 else 1,
  bar_height = if (horizontal) 1 else 5,
  margin = c(1, 1, 1, 1),
  box_adj = c(0.5, 0.5),
  box_fill_col = NA,
  box_border_col = "black",
  bar_border_col = "black",
  bar_lty = 1,
  box_lty = 1,
  bar_lwd = 1,
  box_lwd = 1
)
}
\description{
カラーバーを描画する関数
}
