% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_on_baseplot.R
\name{heatmap_on_baseplot}
\alias{heatmap_on_baseplot}
\title{Heatmap on base plot}
\usage{
heatmap_on_baseplot(
  x,
  row_group = NULL,
  col_group = NULL,
  scale = "none",
  color = RColorBrewer::brewer.pal(11, "RdBu"),
  limits = NULL,
  cell_height = 1,
  cell_width = 1,
  sep_height = 0,
  sep_width = 0,
  groupsep_height = 0.7,
  groupsep_width = 0.7,
  side_rowlab = "r",
  side_rowgrouplab = "l",
  side_collab = "b",
  side_colgrouplab = "t",
  space_rowlab = 0.01,
  space_rowgrouplab = 0.01,
  space_collab = 0.01,
  space_colgrouplab = 0.01,
  args_rowlab = NULL,
  args_rowgrouplab = NULL,
  args_collab = NULL,
  args_colgrouplab = NULL
)
}
\description{
Heatmap on base plot
}
