% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsupporter-package.R
\docType{package}
\name{plotsupporter-package}
\alias{plotsupporter}
\alias{plotsupporter-package}
\title{plotsupporter: A package of convenient functions supporting base plot}
\description{
!!!!!!!!!! TO BE WRITTEN !!!!!!!!!! More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
